wip: remove proof composition
