+++ .vscode/settings.json
@@ -1,6 +1,5 @@
 {
     "rust-analyzer.linkedProjects": [
-        "./light-client-guest/guest/Cargo.toml",
         "./batch-guest/guest/Cargo.toml",
         "./cli/Cargo.toml",
         "./core/Cargo.toml",
