+++ README.md
@@ -4,6 +4,7 @@
 
 Blobstream Zero is an implementation of [Blobstream](https://docs.celestia.org/developers/blobstream) using the [RISC Zero zkVM](https://www.risczero.com/) to verify Celestia blocks.
 
+<!-- TODO update this, outdated -->
 The blocks are verified using a zk Tendermint light client with the [light-client-guest](./light-client-guest/guest/src/main.rs) program, which is composed with a proof that recursively verifies these light client proofs and builds a merkle tree of the verified blocks in [batch-guest](./batch-guest/guest/src/main.rs). This proof is then verified on Ethereum using the [Blobstream solidity contracts](./contracts/src/).
 
 The ABI for the Blobstream Zero solidity contract as well as the merkle tree format for the batch proof is currently matching previous implementations to maintain as much compatibility with previous solutions as well as the [existing APIs to request Blobstream inclusion proofs](https://docs.celestia.org/developers/blobstream-proof-queries#_1-data-root-inclusion-proof).
