+++ contracts/src/ImageID.sol
@@ -19,5 +19,5 @@
 pragma solidity ^0.8.20;
 
 library ImageID {
-    bytes32 public constant BATCH_GUEST_ID = bytes32(0xa661b22447ae53f63d551fbdb0932d9dcaeb226e83e3bb45422e6ec40d3888a4);
+    bytes32 public constant BATCH_GUEST_ID = bytes32(0x78ec222d72105c196e428d11eafb6be20e9a116db361600087310d17c86e06ca);
 }
