+++ core/Cargo.toml
@@ -9,7 +9,6 @@ alloy-sol-types = { version = "0.7", features = ["json"] }
 anyhow = "1.0"
 batch-guest = { path = "../batch-guest" }
 blobstream0-primitives = { path = "../primitives" }
-light-client-guest = { path = "../light-client-guest" }
 risc0-ethereum-contracts = { git = "https://github.com/risc0/risc0-ethereum", tag = "v1.0.0" }
 risc0-zkvm = { version = "=1.0.5" }
 serde_bytes = "0.11"
