--- light-client-guest/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "light-client-guest"
-version = "0.1.0"
-edition = "2021"
-
-[build-dependencies]
-risc0-build = { version = "=1.0.5" }
-
-[package.metadata.risc0]
-methods = ["guest"]
