--- light-client-guest/guest/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "tm-light-client"
-version = "0.1.0"
-edition = "2021"
-
-[workspace]
-
-[dependencies]
-risc0-zkvm = { version = "=1.0.5", default-features = false, features = ['std'] }
-serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-# TODO this is pulling in ed25519 crate unaccelerated, look into if actually used.
-tendermint = { version = "0.39.1", default-features = false }
-tendermint-proto = { version = "0.39.1", default-features = false }
-tendermint-light-client-verifier = { version = "0.39.1", default-features = false, features = [
-	"rust-crypto",
-] }
-blobstream0-primitives = { path = "../../primitives" }
-
-[patch.crates-io]
-sha2 = { git = "https://github.com/risc0/RustCrypto-hashes", tag = "sha2-v0.10.8-risczero.0" }
-# TODO switch to tag when finalized
-ed25519-consensus = { git = "https://github.com/risc0/ed25519-consensus", rev = "8a6d26d1fcd377f9322ba83d3732d04c157e4778" }
-crypto-bigint = { git = "https://github.com/risc0/RustCrypto-crypto-bigint", tag = "v0.5.5-risczero.0" }
