+++ primitives/Cargo.toml
@@ -7,9 +7,6 @@ edition = "2021"
 alloy-sol-types = { version = "0.7", features = ["json"] }
 prost = { version = "0.12", default-features = false }
 prost-types = { version = "0.12", default-features = false }
-serde = { workspace = true }
-serde_bytes = "0.11"
-serde_with = { version = "3.9", default-features = false, features = ["macros", "alloc"] }
 sha2 = { workspace = true }
 tendermint = { workspace = true }
 tendermint-light-client-verifier = { version = "0.39.1", default-features = false, features = [
