+++ primitives/src/lib.rs
@@ -14,8 +14,6 @@
 
 use abi::IBlobstream::DataRootTuple;
 use alloy_sol_types::SolValue;
-use serde::{Deserialize, Serialize};
-use serde_with::{serde_as, Bytes};
 use sha2::Sha256;
 use std::time::Duration;
 use tendermint::merkle::simple_hash_from_byte_vectors;
@@ -54,17 +52,6 @@ pub const DEFAULT_PROVER_OPTS: Options = Options {
     trusting_period: Duration::from_secs(1_209_600),
     clock_drift: Duration::from_secs(0),
 };
-
-#[serde_as]
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LightClientCommit {
-    #[serde_as(as = "Bytes")]
-    pub trusted_block_hash: [u8; 32],
-    #[serde_as(as = "Bytes")]
-    pub next_block_hash: [u8; 32],
-    #[serde_as(as = "Vec<(_, Bytes)>")]
-    pub data_roots: Vec<(u64, [u8; 32])>,
-}
 
 /// Type for the leaves in the [MerkleTree].
 pub type MerkleHash = [u8; 32];
